refactor: create a directory for backend utils library
