+++ apps/drec-api/src/lib/backend-utils/null-or-undefined-result-interceptor.ts
