+++ apps/drec-api/src/lib/backend-utils/success-response.dto.ts
